+++ src/atopile/viewer/src/main.tsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 
-import App from './SchematicApp';
+import App from './App';
 
 import './index.css';
